Update patch set 3
Patch Set 3:

Hi Matthew,

Also we need to set 'worker-reload-mercy' in uwsgi.ini file of the service to some default value. Default value for it is 60 seconds [1].
This value should be dependent on graceful_shutdown_timeout of oslo.service [2] which has default value of 60 seconds. Consider below scenarios:

1. if graceful_shutdown_timeout is set to 0 means service should wait untill all requests are completed. In this case if worker-reload-mercy is not set then after 60 seconds process will be killed even without completing the requests.

2. If graceful_shutdown_timeout is greater than 0 i.e. set to 100 seconds. In this case if worker-reload-mercy is not set then after 60 process will be exit. Ideally process should wait untill 100 seconds before getting killed.

So ideally we should introduce new devstack variable which will calculate the worker-reload-mercy on the basis of graceful_shutdown_timeout and set it to each uwsgi.ini file of the service. For example;

worker_timeout = 90 (60 + 30)

[1] http://uwsgi-docs.readthedocs.io/en/latest/Options.html#worker-reload-mercy
[2] https://docs.openstack.org/oslo.service/latest/configuration/index.html#DEFAULT.graceful_shutdown_timeout

Patch-set: 3