Update patch set 1
Patch Set 1:

An example of the problem that comes from these 502 errors is [1]

(first column is the line-number in the log)

We have the first call come in, then subsquent calls get the 502 error.  But we're already logging things from a second further on in time before we see that.

 3187 23.253.166.156 - - [01/Aug/2017:14:19:47 +0000] "GET /v2.0/auto-allocated-topology/f6806985392e4ece8ac13fb6784131b6 HTTP/1.1" 200 174
 ...
 3186 23.253.166.156 - - [01/Aug/2017:14:20:39 +0000] "GET /identity/v3/auth/tokens HTTP/1.1" 200 3714
 3308 23.253.166.156 - - [01/Aug/2017:14:19:47 +0000] "GET /v2.0/auto-allocated-topology/f6806985392e4ece8ac13fb6784131b6 HTTP/1.1" 502 501
 3309 23.253.166.156 - - [01/Aug/2017:14:20:47 +0000] "GET /identity/v3/auth/tokens HTTP/1.1" 200 3655
 3310 23.253.166.156 - - [01/Aug/2017:14:20:47 +0000] "GET /placement/resource_providers/37ee6cc1-e68f-404a-b9b2-636f9025421b/inventories HTTP/1.1" 200 395
 3311 23.253.166.156 - - [01/Aug/2017:14:19:47 +0000] "GET /v2.0/auto-allocated-topology/f6806985392e4ece8ac13fb6784131b6 HTTP/1.1" 502 501

I'm guessing that happens here is that one worker is processing this, and maybe it gets tied up with the exception and then writes out behind the other worker.  So this might be a red-herring and the real problem remains to be found.  But I'd still prefer to start from a point of consistency with the other CI platforms.

I'm going to propose a change to get better log format in the access log as well.

[1] http://logs.openstack.org/78/480778/2/check/gate-tempest-dsvm-neutron-full-centos-7-nv/8d9e9cc/logs/apache/tls-proxy_access.txt.gz

Patch-set: 1
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0