Remove proxy-sendcl from mod_proxy_uwsgi apache path
Calling setenv appears to be globally scoped which is breaking the
glance path which relies on chunked uploads. The glance path is
separated by using mod_proxy instead of mod_proxy_uwsgi because
mod_proxy_uwsgi doesn't support chunked encoding.[1] The proxy-sendcl [2]
was set on the mod_proxy_uwsgi path just in case someone tried to send a
chunked request to the api server we would be able to handle it. It
tells apache to locally cache the chunked request and send the
content-length as a normal upload to the upstream server. However, if we
can only set it globally across then small potential benefit is not worth
having all glance uploads cached by apache. This commit just removes
setting the flag. In the future if we can have devstack isolate this
flag it might be worth adding back to the mod_proxy_uwsgi path, but for
right now it's not worth the tradeoff.

As part of this there were some misconfigurations in apache and uwsgi
around handling chunked encoding. This was overlooked because we weren't
actually chunking it over the proxy connection to glance/uwsgi. When we
stopped setting sendcl in apache things stopped working because uwsgi
didn't know how to handle the data.

[1] https://github.com/unbit/uwsgi/issues/1540
[2] https://httpd.apache.org/docs/2.4/mod/mod_proxy.html#request-bodies

Change-Id: Iab2e2848877fa1497008d18c05b0154892941589
Closes-Bug: #1709970