Update patch set 2
Patch Set 2:

So just an update on this, I've done some testing and I made an assumption here on rev 2 that this works because when I did the glance upload manually against the uwsgi endpoint (without apache) it returned the 204. But what was really happening is uwsgi was dropping the chunked data body and glance was writing a 0 sized image. The 502s we're seeing here are likely being caused by apache understanding the upstream http server (uwsgi in this case) isn't doing the right thing and bailing out.

I'm at a loss right now on how to get this working. The tcpdump against just uwsgi is quite weird looking, so I'm thinking this might be a uwsgi bug. But I still need to do some more investigation.

Patch-set: 2
Label: Verified=0