Update patch set 2
Patch Set 2:

For reference mtreinish and I did some talking about this in IRC: http://p.anticdent.org/32fX

My guess is that the interface between whatever is reading chunked uploads and uwsgi is not working well, in part because of http://uwsgi-docs.readthedocs.io/en/latest/Chunked.html

I looked in the glance code to try and figure out where this was, but my exhausted Friday evening brain got nowhere. Glance is indirect, to say the least.

Patch-set: 2