Update patch set 3
Patch Set 3:

The latest rev adds a depends on the glance patch adding the uwsgi.chunked_read() api usage to the glance paste pipeline. The only place where I've seen a working chunked request go through is in the eventlet case. In eventlet it reads the chunked data before passing it to the wsgi app. (see the link in the depends on glance patch) So I just did that via a middleware app if uwsgi is detected.

Patch-set: 3
Label: Verified=0