Don't special case Grenade + systemd
Now that the pike branch exists we don't want to special case grenade +
system on the target side. We should use systemd for both sides of the
pike -> master upgrade. Note this change should not be backported so
that we do not attempt to use systemd on the ocata -> pike upgrade path.

Change-Id: I6198bf1842a44773fce80672c81eee3afc3c6f38