Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

It seems like we've seen this before because Ib41319676d2fd24144a1493bd58543ad71eb8d6c changed things to only allow alpha-numeric values.

But this isn't checked for the localrc case.  So I'd just like to see us run the same check, and give the same error, if the passwords are coming from config file.  That seems like it will save everyone a bunch of time debugging this in the future.

Patch-set: 1
Label: Code-Review=-1