Update patch set 2
Patch Set 2:

> doing computes in parallel is really not supported, the control
 > stack needs to be up first.

so then what's the point of even having a 60s hard-coded wait if
the control needs to be up first?

we have been doing things in parallel for some time now and it's
worked great.

if this is a non-starter patch, we can find a hack in our jobs to
poll on our control node before kicking off the compute stack.sh

Patch-set: 2