Update patch set 2
Patch Set 2:

> doing computes in parallel is really not supported, the control
 > stack needs to be up first.

The 60s poll is arbitrary and not guaranteed to work so there should be a config option to set the value accordingly. This applies regardless if it is running in parallel is supported or not.

I do agree with Jamo, though - this is something that has always worked so we should attempt to support it going forward.

Patch-set: 2
Reviewer: Gerrit User 11240 <11240@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0