Add etcd3 support for s390x
Devstack fails on s390x, as there is no etcd3 binary hosted
on github like it is for other architectures.

The current workaround for s390x is to disable etcd3 in local.conf.
But there might be situations where etcd3 is required. This patch
allows the installation of etcd3 for s390x if an alternate download
URL is provided via local.conf

  ETCD_DOWNLOAD_URL=<custom-download-url>

If this is not the case, devstack will fail like before if etcd3
service is enabled.

In the long run the issue gets solved once distros provide the etcd3 package.

[1] https://github.com/openstack-dev/devstack/commit/d8bb220606737719bcdf7c5b4f54906f2974c71c

Change-Id: I1c378a0456dcf2e94d79a02de9d3e16753d946d6
Partial-Bug: #1693192