Only do etcd3 config if service is enabled
Devstack is setting etcd3 vars even though if the etcd3 service disabled.
This causes devstack on s390x to fail as no etcd3 binary for s390x is
available. The workaround till [1] was to just disable the etcd3 service
in local.conf [2][3]. This patch is enabling this workaround again.

[1] https://github.com/openstack-dev/devstack/commit/87943ab88e760598710952598424e7feded88f5c
[2] https://bugs.launchpad.net/devstack/+bug/1693192
[3] https://docs.openstack.org/devstack/latest/configuration.html#kvm-on-s390x-ibm-z-systems

Change-Id: Ifcbc8b990e853a73462fa36590c24e210967e03a
Partial-Bug: #1693192