Update patch set 7
Patch Set 7:

After a discussion with Sean I came to the conclusion that this is the proper way of doing it. ETCD3 is a base service that should always be available - so let's go this path. For the user that means that he needs to provide a binary on its own. that's ok for now. In the long run I need to work on getting a binary hosted somewhere... Thanks!

Patch-set: 7