WIP: Fix running with SERVICE_IP_VERSION=6
There are some locations where we need the raw IPv6 address instead of the
url-quoted version enclosed in brackets.

Also fix generation of TLS certs not to include the IP address as DNS SAN, in
order to avoid an urllib3 bug [1].

Change default SERVICE_IP_VERSION=6 for testing, will revert that part once
tests turn out to be passing.

[1] https://github.com/shazow/urllib3/issues/1269

Change-Id: Id074be38ee95754e88b7219de7d9beb06f796fad
Closes-Bug: 1656329