Update patch set 6
Patch Set 6:

Changing SERVICE_HOST to be the un-quoted IP address and adding a function that adds the [] when needed might be possible, but it would be a much larger patch and it would also be an incompatible change for all plugins or other projects (neutron-* most likely) that might be tuned to the current behaviour.

So maybe a function that removes the [] on demand?

Patch-set: 6
Label: Verified=0