Update patch set 2
Patch Set 2:

This shouldn't be necessary and it would break anyone who was setting OS_AUTH_URL with a versioned endpoint. keystoneauth does version discovery and will use the right endpoint for the given OS_IDENTITY_API_VERSION. The unversioned endpoint works fine for other clients.

Also, not that this is very relevant with v2 being ripped out, but OS_IDENTITY_API_VERSION can be a 2 or a 3 but the v2 endpoint is at /v2.0 not /v2, so this would be incorrect if we still supported keystone v2.0.

Patch-set: 2
Reviewer: Gerrit User 8482 <8482@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Workflow=0