Update patch set 2
Patch Set 2:

Sean- it's a fair concern. For now I'm going to ask that we leave them top level. It's not important for playbooks - but it IS important for roles, as roles defined here can be used by other jobs and it's important for them to be in a directory called 'roles' for that to work.

Let's not lose track of that concern though, and come back to it once the dust settles a bit.

Patch-set: 2
Reviewer: Gerrit User 2 <2@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0