Update patch set 4
Patch Set 4:

All legacy jobs here fail because the look for setup_host in d-g, which has been removed due to the patch in depends-on. 

We would have to update legacy jobs as well, but we would run in a circular dependency of patches. I think we need to remove the depends-on from here.

It should be fine since zuul was not able to honour the depends-on when checking for name conflicts anyways.

Patch-set: 4
Label: Verified=0