Enable arp_responder with DVR
arp_responder should be enabled by default in DVR installation.
Also when arp_responder is enabled, DVR HA migration tempest scenario
tests are consistently passing(Please see my comments in [1]).
But with linuxbridge agent there is a known issue [2] with l2pop
and arp_responder, so we are enabling it for only openvswitch agent.

[1] https://review.openstack.org/#/c/500384/
[2] https://bugs.launchpad.net/neutron/+bug/1661717

Related-Bug: #1714802
Change-Id: I1040e34a4ba69d9fe3aca585e4fe0649515c1730