Force pip install update of PasteDeploy for Ceph
The gate-tempest-dsvm-full-devstack-plugin-ceph-ubuntu-xenial job is
currently failing at a rate of 100% since we updated to use the Pike
UCA. The Ceph packages in the Pike UCA are pulling in python-paste and
python-pastedeploy. When devstack pip installs a newer version of
python-paste to satisfy upper-constraints, we end up with Paste
installed from pypi and python-pastedeploy installed from the Ubuntu
archive and Keystone fails to start with the error:

  "ImportError: cannot import name deploy."

This adds a hack to force an update of PasteDeploy from pypi to work
around the problem. I had wanted to add something to the post-config
step in the devstack-plugin-ceph but learned recently that post-config
now runs _after_ Keystone has started. The end of the install phase is
too early too because that occurs _before_ devstack pip installs
packages. So we really need a new phase that happens before all of the
services start but after all of the pip installs complete, that plugins
can use.

Change-Id: I7eebf7e9d4da9732c3dc65847d3a65477e7e641c