Update patch set 3
Patch Set 3: Workflow+1 Code-Review+2

The dual python install path is part of what makes this extra fragile.

I'd honestly really like to delete that, we should pick one python version at start time and operate on that exclusively. The whole dual path basically exists for swift to still be run in python3 jobs, under the assumption that it would be short lived need. But as that doesn't seem to be the case, I think the kinds of bugs that show up in the infrastructure because of the fragility here don't warrant it.

Patch-set: 3
Reviewer: Gerrit User 2750 <2750@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2
Label: Workflow=+1