See if we can speed up creating projects/users/roles
DevStack has not kept up with OSC abilities to a) not required IDs for
every last thing, and b) cleanly not error out when creating resources.

Let's try batching all of the keystone setup calls to OSC into a single
call; errors here are fatal anyway so we don't risk much.

Change-Id: I430aaeaecbab3e07d85ab331da727f3f6433feb8