Update patch set 6
Patch Set 6:

So I think the issue here is conductor is hitting the rpc_response_timeout and this decorator code in nova is automatically retrying:

http://logs.openstack.org/18/507918/6/check/gate-tempest-dsvm-neutron-full-ubuntu-xenial/d4f175d/logs/screen-n-super-cond.txt.gz#_Oct_04_18_32_07_052398

https://github.com/openstack/nova/blob/353db2d1932965b6502e002b8be510440ff529c0/nova/scheduler/utils.py#L599

It will do that up to max_attempts-1, so that's why we have 2 retries because max_attempts defaults to 3.

That's probably a bad idea given we're claiming in the scheduler now. We could be allocating too many resources for a single instance.

Patch-set: 6
Label: Verified=0