Update patch set 4
Patch Set 4:

> Instead of this, wouldn't it be better to put these packages into
 > files/*/ceph
 > 
 > these will be installed first.  then the pip install later will
 > overwrite

The problem here is that the Ceph packages in the Pike UCA pull in python-paste and python-pastedeploy, but the latest python-paste version doesn't satisfy upper-constraints. So python-paste gets overwritten by pip (to install a newer version from pypi) but python-pastedeploy does not. So we end up with a mismatch.

Normally I would want to take care of this in a devstack phase that occurs AFTER pip install but BEFORE keystone is started (keystone start is failing for the ceph job only because of "ImportError: cannot import name deploy"). But we no longer have a phase like that. The post-config phase runs AFTER keystone is started.

So we need to force a pip install update of PasteDeploy so that both are installed in the same location.

Does that make sense?

Patch-set: 4