Fix etcd advertise and listen client URLs
Theses URLs are currently set to SERVICE_HOST, which makes etcd
unable to start in subnode. In order to fix it, this patch changes
the following URLs:
* advertise-client-urls: List of this member's client URLs to
  advertise to the rest of the cluster. These URLs can contain domain
  names [1]. Example: "http://example.com:2379, http://10.0.0.1:2379"
* listen-client-urls: List of URLs to listen on for client traffic.
  This flag tells the etcd to accept incoming requests from the clients
  on the specified scheme://IP:port combinations. Scheme can be either
  http or https. If 0.0.0.0 is specified as the IP, etcd listens to
  the given port on all interfaces. If an IP address is given as well
  as a port, etcd will listen on the given port and interface. Multiple
  URLs may be used to specify a number of addresses and ports to
  listen on. The etcd will respond to requests from any of the listed
  addresses and ports [2].

In before, these two URLs were set by using SERVICE_HOST, which
seems incorrect, because SERVICE_HOST is not necessary the IP address
of current host. However, etcd have to advertise or listen to
an IP address/domain of current host. Use HOST_IP for these URLs
should be a more reasonable choice.

[1] https://coreos.com/etcd/docs/latest/v2/configuration.html
    #advertise-client-urls
[2] https://coreos.com/etcd/docs/latest/v2/configuration.html
    #listen-client-urls

Change-Id: I27b29349d8d55c588f88461ac3658fd1b1342a1f
Closes-Bug: #1720240