Update patch set 2
Patch Set 2:

> Neutron has been fixed, I still think other projects should fix
 > their jobs, like I said before:
 > 
 > > The bug comes from kuryr-libnetwork enabling etcd on the subnode,
 > which is wrong IMO.
 > 
 > Reverting this would also break https://review.openstack.org/505168.

I think the fix should be on devstack since it seems misconfigure etcd3 which causes the failure. If a full revert is undesirable, an alternative is to do a partial revert: https://review.openstack.org/#/c/511092/ . The bottom is that the listen-client-urls and advertise-client-urls should be set as suggested by the etcd documentation.

Patch-set: 2