Update patch set 2
Patch Set 2:

Thanks for working on this!

It would be nice if this used Id45a50821539985710da7172a9185fd91e803603 since that's going to be the generic mechanism to stage files on the remote host and then pull everything to the executor and push to log server (for longs and configs).

But I understand if you prefer not to wait for that patch to merge and just go ahead with this.

Patch-set: 2
Reviewer: Gerrit User 1921 <1921@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0