Checkout from -eol tag if stable branch does not exist
Since cinder and neutron's stable/newton branches have been
end of life'd, we can't install base devstack on stable/newton.

This modifies the git_clone function to check, as a last resort,
if the git_ref is a stable branch and if there is a corresponding
-eol tag, to use that for the checkout rather than the stable branch.

This is needed in the cases that projects are end of life'd out of
order, as is the case with neutron and cinder being end of life before
nova for stable/newton.

Co-Authored-By: Eric Fried <efried@us.ibm.com>

Change-Id: Ic2a98ef4933b9b9fc05cd5eb67cc89a86b2e5292