Update patch set 2
Patch Set 2: Code-Review-1

Why would they not share the same service?  afaik the apiserver can be setup to use it's own prefix so i don't see a conflict there [1]?  it's not clear if this is related to writing a plugin or what, can you explain why this is necessary?

[1] https://kubernetes.io/docs/reference/generated/kube-apiserver/

Patch-set: 2
Label: Code-Review=-1