Set the base set of services
Use the test-matrix role from devstack-gate to define a base set of services
to be enabled for the controller and compute nodes.

Extend the local conf module to handle the base set of services.

Since the test-matrix defines services for primary and subnode nodes, we
need a multinode job to test that this works. Add a new host group called
subnode that includes the non-controller hosts. Add a new job that runs
devstack on a two nodes environment.

Depends-on: Ie36ba0cd7cfcd450b75000a76a64d856f2a83eba
Change-Id: I379abf482c89122533324e64fefbff3d5a618a89