Update patch set 1
Patch Set 1:

Hi, Bob. If we set "DEST"(for example: /opt/stack/openstack) in local.conf to specify the install destination of the openstack, there would be an error in install_plugin because scripts can't find os-xenapi in /opt/stack/openstack folder. Add this flag could help to download and locate os-xenpai in the folder

Patch-set: 1