Use systemctl restart in _run_under_systemd
'systemctl restart' means either restart or start, but 'start' means
run if not running, otherwise return okay.

This presents a problem for some services when they were were running
under systemd in the OLD target in grenade: they do not restart.

See the linked bug for more detail, and links to examples of the
problem.

Change-Id: I6e2a3bf4f0a66292d5d1aede509f43eb90a720ae
Closes-Bug: #1736385