Use stage_dir instead of /opt/stack
Historically we have collected devstack logs under /opt/stack.
Stop doing that and collect them in the stage_dir instead, so that
once the base job logs pull service comes around we are ready for it.

This add the benefit of writing things into a folder which is
already owned by the ansible user (ansible_user_dir), so we don't
run into issue writing there.

Change-Id: Ib4be2f5056c0dc2b776de4a0d18b47b12624be92