Update patch set 1
Patch Set 1: Code-Review-1

Thanks for pointing this out, I'm glad this is the conclusion that's been reached here.

I think we need to do some untangling, however.  one significant issue is that python2-pip does *not* have this behaviour, only pip3 (i just tested this installing a simple pkg with pip & pip3 and you end up with two binaries, the python2 one in /usr/bin/bashate & the python3 one in /usr/local/bin/bashate)

---
 [root@f27-test bashate]# whereis bashate
 bashate: /usr/bin/bashate /usr/local/bin/bashate
---

this is further complicated by the gate, where we have traditionally overwritten the system packaged pip & setuptools during build as we often required the latest to workaround issues not addressed in packaged versions [1].

so we're going to have to figure this so it works for python 2 & 3, and inside & outside the gate.

i think the python2/3 split is the first thing to address in this...

we should have gate nodes for fedora27 soon, i just released dib 2.10.0 which can build them a few hours ago

[1] https://git.openstack.org/cgit/openstack/diskimage-builder/tree/diskimage_builder/elements/pip-and-virtualenv

Patch-set: 1
Label: Code-Review=-1