Update patch set 2
Patch Set 2:

I fixed my change to only change the behaviour on Python 3. The CI tests no pass. Would you mind to review this new change please?

Ian Wienand: "so we're going to have to figure this so it works for python 2 & 3, and inside & outside the gate."

If your comment is about my change: I tested it manually with Python 2 and Python 3 on DevStack, the computed exec_prefix is correct.

Moreover, tests passed on the gate on Python 2 and Python 3:

* legacy-grenade-dsvm-neutron: SUCCESS
* legacy-tempest-dsvm-py35: SUCCESS

Patch-set: 2