Update patch set 2
Patch Set 2:

I think i'm ok with this, but note that none of the gate tests have run on fedora.  we are close to having fedora27 nodes then we can bring up an experimental fedora27 job; unfortunately there was a regression with dib 2.10 (which has f27 support) and we had to revert images, but the 2.10.1 point release i did today should be good.

Patch-set: 2