Update patch set 3
Patch Set 3:

Ghanshyam,

This is exactly the same you do in https://review.openstack.org/#/c/531195/ .

https://review.openstack.org/#/c/525038/1/.zuul.yaml was wrong since it would *only* run the job if roles, playbooks gets changed.

Regarding your two items:
Ad 2): Any other repo can inherit from this one or add to the job directly in the project section its own irrelevant-files list and that overrides this one. This list is the default - and if you look at the referenced followup (see needed-by), it will make project-config much simpler.

Ad 1) Yes, mixing files and irrelevant-files is a challenge. You can have both, it's not an error per se but this is not tested and might open some interesting effects.

See the followup - do you have a better idea to simplify this?

Patch-set: 3