Update patch set 3
Patch Set 3: Code-Review+1

- https://review.openstack.org/#/c/531195/ this is integrated job not very base job and hardly anyone override from this.

>Ad 2): .... it will make project-config much simpler.

gmann - +1 on this and i agree the benefit of this with overriding the irrelevant-files if needed. 

>Ad 1) Yes, mixing files and irrelevant-files is a challenge. You >can have both, it's not an error per se but this is not tested >and might open some interesting effects.
>See the followup - do you have a better idea to simplify this?

gmann- yea this is tricky and we do not know the behavior. 1 idea is to change in zuul logic - Give Preference 'files' over irrelevant-files if both are present. 

m +1 on this as it is not error and we can handle when we have any side effect.

>See the followup
you have any followup patch fir testing the 1st case?

Patch-set: 3
Reviewer: Gerrit User 8556 <8556@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1