Remove 'python_version', 'python3_version' functions
These look like they should have been used starting with the original
change [1], but they were not. Instead, there's only two calls to one of
the functions, and the other function is uncalled and contains a typo.
Update the two calling function to use the PYTHON variable instead and
just remove these functions.

[1] Id826f275b99b9f397b95e817941019fc503daa1d

Change-Id: I92fb3eeb5a111c4c5a55ac4293611fa308e9ee11