Redirect stack.sh output to devstack_early_log
In the run devstack role we specify a path to devstack_early_log then
hand it to stack.sh as a parameter which stack.sh does nothing with. We
need to actually redirect the output of stack.sh into this file (this is
what devstack-gate did with it at least). Add the redirect into the
devstack_early_log path here.

Change-Id: I9bedbe91c60257d94173b1c70676dd6c2b49dc91