Update patch set 3
Patch Set 3: Code-Review-1

As discussed on IRC, I think that this is not the right approach. 
Devstack introduced plugins so that a service that it's not configured by the core devstack can be *totally* configured by its Devstack plugin.

So I think that the list should only contain the services directly configured by devstack. Every other service, configured through a devstack plugin, should:
a) default the value of service_available.<service> to False; and
b) set the value of service_available.<service> to True in the relevant step (test-config) of the devstack plugin.

For example in Sahara we implement b):
http://git.openstack.org/cgit/openstack/sahara/tree/devstack/plugin.sh#n252

and I'm going to file a patch which fixes a).

-1 because at least sahara should not be in the list, if not all other services.

Patch-set: 3
Reviewer: Gerrit User 10459 <10459@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1