Update patch set 3
Patch Set 3:

@Luigi Toscano 

>a) default the value of service_available.<service> to False; and
It is always true as default on each plugin side. This config option is register on tempest plugin side not by service. 

regarding doing these on devstack plugin side, we can do that but that did not happen and leads to this bug. I do not think its too bad to set the tempest required service in common place o0n devstack side.

Patch-set: 3