Update patch set 3
Patch Set 3:

one more point that why it would not work when setting on devstack plugin side. If service is not setup which means their devstack plugin will not set the service availability on tempest and it stay as True because every plugin have it default to True.

We can make default value to False but that backward incompatible change we need that with upgrade notes, that seems not worth for me for small benefit.

Patch-set: 3