[WIP][DNR] Work around with Devstack multinode gate problem
Multinode job is currently not working. It has several problems which
relate to devstack-gate & devstack procedure. In the legacy multinode gate
setup, devstack-gate writes some more variables in subnode's localrc_file
[1]. These variables will be used in devstack, for example, database [2].
In the Zuulv3 native gate, subnode doesn't get these variables so devstack
mutlinode gate failed.

This patch is just for work around with this problem. We have to change
the procedure if we want to do actual fix.

[1] https://github.com/openstack-infra/devstack-gate/blob/master/devstack-vm-gate.sh#L567-L588
[2] https://github.com/openstack-dev/devstack/blob/master/lib/database

Change-Id: Iae41d6f7baf6f0f781be4eb47a06a64bf66f8dae