Update patch set 3
Patch Set 3: Code-Review+2

As described in the other bug, this started because Ia14bfa6c89fd04844842d727d11ff8874ec7fc2a merged which brings this into nova.  as a regex library, it seems to be the type of thing that gets dragged in as a pretty general dependency, so i think i'm ok with it here.  we should probably follow-on and remove this from the horizon list though

Patch-set: 3
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2