Update patch set 5
Patch Set 5: Code-Review+2

As everyone has said, a static key does set of everyone's "spidey sense" of this could be a problem :)  

I'm sort of +1.5 on this ... if we want a random per-test global key, we could create one in the ansible playbooks (or devstack-gate, but we're trying to not add too many more features).  Was that considered?

Patch-set: 5
Reviewer: Gerrit User 7118 <7118@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2