Update patch set 2
Patch Set 2: Code-Review-1

(1 comment)

I agree this is wrong as is ... but ...

The logs this produces [1] aren't really what was intended for devstack-early; it's supposed to be capturing the stdout and tracing output before before we start all the fd magic to send things via log files.  [2] is actually a good example, because you'll note that we don't actually have that failure output anywhere else in the other logs of [3].

In the old days, this was capturing the output of "ansible-playbook -m shell ... stack.sh" which otherwise disappeared; now this is a task inside a bigger playbook the output is captured anyway.

You can see that in [4] where we actually have the early output captured as part of the ansible job-output.txt

Thus, I think we want to get rid of this all together

[1] http://logs.openstack.org/79/535879/2/check/devstack/d94bede/controller/logs/devstack-early.txt.gz
[2] http://logs.openstack.org/86/539686/3/check/legacy-tempest-dsvm-neutron-full-centos-7/6617c49/logs/devstack-early.txt.gz
[3] http://logs.openstack.org/86/539686/3/check/legacy-tempest-dsvm-neutron-full-centos-7/6617c49/logs/
[4] http://logs.openstack.org/79/535879/2/check/devstack/d94bede/job-output.txt.gz#_2018-02-02_22_14_14_418227

Patch-set: 2
Label: Code-Review=-1