Update patch set 6
Patch Set 6:

The devstack log does not show everything that is done by devstack, which was the reason always for having the devstack-early file. The port I made of that in ansible land does not work, but I think it would be worth fixing it and take the devstack log out of the job-status.

Having the devstack log in job-status may be convenient when streaming logs, but how's that going to work for multinode?

Patch-set: 6